Notice refactor -- Use type inference instead of visitor pattern for addNotice methods
